Implementado a conexão com o banco de dados e ajustando metodos

+++ RestApplication/RestApplication/Services/IPersonService.cs
@@ -5,9 +5,9 @@ namespace RestApplication.Services
     public interface IPersonService
     {
         Person Create (Person person);
-        Person FindById(long id);
+        Person FindById(int id);
         List<Person> FindAll();
         Person Update(Person person);
-        void Delete(long id);
+        void Delete(int id);
     }
 }
